requirements.txt
# Wazo Receptionniste IA - Dependencies

# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Asterisk ARI
ari-py==0.1.3
websockets==12.0

# OpenAI
openai==1.12.0

# Audio processing
pydub==0.25.1

# HTTP client
httpx==0.26.0
aiohttp==3.9.3

# Environment
python-dotenv==1.0.1

# Logging
structlog==24.1.0
